+++ components/lsp-form/LSPForm.tsx
@@ -1,5 +1,4 @@
 import React, { FC, useState, useEffect } from "react";
-import { DateTime, Duration } from "luxon";
 import Tabs from "../tabs";
 import {
   Wrapper,
@@ -7,10 +6,12 @@ import {
   SettleWrapper,
   SettleTitle,
   TimeRemaining,
+  SettleText,
 } from "./LSPForm.styled";
 
 import MintForm from "./MintForm";
 import RedeemForm from "./RedeemForm";
+import { calculateTimeRemaining } from "./helpers";
 
 const LSPForm: FC = () => {
   const [showSettle, setShowSettle] = useState(false);
@@ -43,6 +44,10 @@ const LSPForm: FC = () => {
         <SettleWrapper>
           <SettleTitle>Settle Position</SettleTitle>
           <TimeRemaining>Time Remaining: {timeRemaining}</TimeRemaining>
+          <SettleText>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer
+            placerat malesuada sapien ut dapibus. Aliquam.
+          </SettleText>
           <SettleButton onClick={() => setShowSettle(false)}>
             Settle
           </SettleButton>
@@ -53,15 +58,3 @@ const LSPForm: FC = () => {
 };
 
 export default LSPForm;
-
-// Stub function to show effect.
-const calculateTimeRemaining = () => {
-  const utc = DateTime.local().toUTC().endOf("hour").toMillis();
-  const difference = utc - DateTime.local().toMillis();
-
-  let text = "00:00";
-  // format difference
-  if (difference > 0) text = Duration.fromMillis(difference).toFormat("mm:ss");
-
-  return text;
-};
